remove redundant test
The removed test was intended to check that a `terraform apply` would result in error if an attribute marked as Required in the schema was missing. No custom function behaviour had been implemented for this attribute. It is not necessary to test primitive behaviours such as Required in provider acceptance tests, as this is the responsibility of Terraform Core.